feat: Ticket에 수동로또 개수를 받는 생성자 추가
- rename Ticket -> PurchaseTicket
- 수동 로또 개수를 검증하는 로직 추가
